+++ backend/src/main/java/com/example/courtierprobackend/audit/loginaudit/dataaccesslayer/LoginAuditEvent.java
@@ -1,4 +1,4 @@
-package com.example.courtierprobackend.audit.dataaccesslayer;
+package com.example.courtierprobackend.audit.loginaudit.dataaccesslayer;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -6,17 +6,13 @@ import lombok.*;
 import java.time.Instant;
 import java.util.UUID;
 
-/**
- * Entity to track password reset events for audit purposes.
- * Records when users request password resets and when they complete them.
- */
 @Entity
-@Table(name = "password_reset_events")
+@Table(name = "login_audit_events")
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PasswordResetEvent {
+public class LoginAuditEvent {
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
@@ -29,8 +25,7 @@ public class PasswordResetEvent {
     private String email;
 
     @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private ResetEventType eventType; // REQUESTED or COMPLETED
+    private String role;
 
     @Column(nullable = false)
     private Instant timestamp;
@@ -40,9 +35,4 @@ public class PasswordResetEvent {
 
     @Column(length = 500)
     private String userAgent;
-
-    public enum ResetEventType {
-        REQUESTED,  // User clicked "Forgot Password" and email was sent
-        COMPLETED   // User successfully reset their password
-    }
 }
