+++ backend/src/main/java/com/example/courtierprobackend/audit/loginaudit/dataaccesslayer/LoginAuditEventRepository.java
@@ -1,4 +1,4 @@
-package com.example.courtierprobackend.audit.dataaccesslayer;
+package com.example.courtierprobackend.audit.loginaudit.dataaccesslayer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
