+++ backend/src/main/java/com/example/courtierprobackend/audit/passwordresetaudit/businesslayer/PasswordResetAuditService.java
@@ -1,7 +1,7 @@
-package com.example.courtierprobackend.audit.businesslayer;
+package com.example.courtierprobackend.audit.passwordresetaudit.businesslayer;
 
-import com.example.courtierprobackend.audit.dataaccesslayer.PasswordResetEvent;
-import com.example.courtierprobackend.audit.dataaccesslayer.PasswordResetEventRepository;
+import com.example.courtierprobackend.audit.passwordresetaudit.dataaccesslayer.PasswordResetEvent;
+import com.example.courtierprobackend.audit.passwordresetaudit.dataaccesslayer.PasswordResetEventRepository;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
