+++ backend/src/main/java/com/example/courtierprobackend/audit/passwordresetaudit/dataaccesslayer/PasswordResetEvent.java
@@ -1,4 +1,4 @@
-package com.example.courtierprobackend.audit.dataaccesslayer;
+package com.example.courtierprobackend.audit.passwordresetaudit.dataaccesslayer;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -6,13 +6,17 @@ import lombok.*;
 import java.time.Instant;
 import java.util.UUID;
 
+/**
+ * Entity to track password reset events for audit purposes.
+ * Records when users request password resets and when they complete them.
+ */
 @Entity
-@Table(name = "login_audit_events")
+@Table(name = "password_reset_events")
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginAuditEvent {
+public class PasswordResetEvent {
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
@@ -25,7 +29,8 @@ public class LoginAuditEvent {
     private String email;
 
     @Column(nullable = false)
-    private String role;
+    @Enumerated(EnumType.STRING)
+    private ResetEventType eventType; // REQUESTED or COMPLETED
 
     @Column(nullable = false)
     private Instant timestamp;
@@ -35,4 +40,9 @@ public class LoginAuditEvent {
 
     @Column(length = 500)
     private String userAgent;
+
+    public enum ResetEventType {
+        REQUESTED,  // User clicked "Forgot Password" and email was sent
+        COMPLETED   // User successfully reset their password
+    }
 }
