+++ backend/src/main/java/com/example/courtierprobackend/security/AuthenticationEventListener.java
@@ -1,6 +1,6 @@
 package com.example.courtierprobackend.security;
 
-import com.example.courtierprobackend.audit.businesslayer.LoginAuditService;
+import com.example.courtierprobackend.audit.loginaudit.businesslayer.LoginAuditService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
