+++ backend/src/test/java/com/example/courtierprobackend/security/AuthenticationEventListenerTest.java
@@ -1,6 +1,6 @@
 package com.example.courtierprobackend.security;
 
-import com.example.courtierprobackend.audit.businesslayer.LoginAuditService;
+import com.example.courtierprobackend.audit.loginaudit.businesslayer.LoginAuditService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
